requirements.txt
streamlit>=1.39.0
langgraph>=0.2.50
langchain-openai>=0.2.10
langchain>=0.3.10
python-dotenv>=1.0.0
langsmith>=0.3.0 